axi_dmac: Implement generic transfer start sync
This commit adds a generic 'sync' port used for transfer start.
The signal used for synchronization is assumed to be synchronous
with the interface clock.

On the receive AXI Streaming interface, the user can choose to
receive the synchronization signal on TUSER instead of 'sync',
which is the default setting, configurable through the
AXIS_TUSER_SYNC synthesis parameter.

Signed-off-by: Ionut Podgoreanu <ionut.podgoreanu@analog.com>